chore: Move more functions to gallia.net
